+++ packages/device_info_plus/device_info_plus/example/integration_test/device_info_plus_test.dart
@@ -76,16 +76,18 @@ void main() {
     expect(iosInfo.utsname.version, isNotNull);
   }, skip: !Platform.isIOS);
 
-  testWidgets('Check all android info values are set',
+  testWidgets('Check all android info values are available',
       (WidgetTester tester) async {
-    expect(androidInfo.version.baseOS, isNotNull);
+    if (androidInfo.version.sdkInt >= 23) {
+      expect(androidInfo.version.baseOS, isNotNull);
+      expect(androidInfo.version.previewSdkInt, isNotNull);
+      expect(androidInfo.version.securityPatch, isNotNull);
+    }
+
     expect(androidInfo.version.codename, isNotNull);
     expect(androidInfo.version.incremental, isNotNull);
-    expect(androidInfo.version.previewSdkInt, isNotNull);
     expect(androidInfo.version.release, isNotNull);
-    expect(androidInfo.version.sdkInt, equals(30));
-    expect(androidInfo.version.securityPatch, isNotNull);
-
+    expect(androidInfo.version.sdkInt, isNotNull);
     expect(androidInfo.board, isNotNull);
     expect(androidInfo.bootloader, isNotNull);
     expect(androidInfo.brand, isNotNull);
@@ -93,6 +95,11 @@ void main() {
     expect(androidInfo.display, isNotNull);
     expect(androidInfo.fingerprint, isNotNull);
     expect(androidInfo.hardware, isNotNull);
+
+    expect(androidInfo.displayMetrics.heightPx, isNotNull);
+    expect(androidInfo.displayMetrics.widthPx, isNotNull);
+    expect(androidInfo.displayMetrics.yDpi, isNotNull);
+    expect(androidInfo.displayMetrics.xDpi, isNotNull);
 
     expect(androidInfo.host, isNotNull);
     expect(androidInfo.id, isNotNull);
@@ -110,7 +117,7 @@ void main() {
     expect(androidInfo.systemFeatures, isNotNull);
   }, skip: !Platform.isAndroid);
 
-  testWidgets('Check all macos info values are set',
+  testWidgets('Check all macos info values are available',
       ((WidgetTester tester) async {
     expect(macosInfo.computerName, isNotNull);
     expect(macosInfo.hostName, isNotNull);
@@ -137,4 +144,97 @@ void main() {
     expect(linuxInfo.variant, isNull);
     expect(linuxInfo.variantId, isNull);
   }), skip: !Platform.isLinux);
+
+  testWidgets('Check all Windows info values are available',
+      ((WidgetTester tester) async {
+    expect(
+      windowsInfo.numberOfCores,
+      isPositive,
+    );
+    expect(
+      windowsInfo.computerName,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.systemMemoryInMegabytes,
+      isPositive,
+    );
+    expect(
+      windowsInfo.userName,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.majorVersion,
+      equals(10),
+    );
+    expect(
+      windowsInfo.minorVersion,
+      equals(0),
+    );
+    expect(
+      windowsInfo.buildNumber,
+      greaterThan(10240),
+    );
+    expect(
+      windowsInfo.platformId,
+      equals(2),
+    );
+    expect(
+      windowsInfo.reserved,
+      isZero,
+    );
+    expect(
+      windowsInfo.buildLab,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.buildLab,
+      startsWith(
+        windowsInfo.buildNumber.toString(),
+      ),
+    );
+    expect(
+      windowsInfo.buildLabEx,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.buildLab,
+      startsWith(windowsInfo.buildNumber.toString()),
+    );
+    expect(
+      windowsInfo.digitalProductId,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.editionId,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.productId,
+      isNotEmpty,
+    );
+    expect(
+      RegExp(r'^([A-Z0-9]{5}-){4}[A-Z0-9]{5}$')
+              .hasMatch(windowsInfo.productId) ||
+          RegExp(r'^([A-Z0-9]{5}-){3}[A-Z0-9]{5}$')
+              .hasMatch(windowsInfo.productId),
+      isTrue,
+    );
+    expect(
+      windowsInfo.productName,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.productName,
+      startsWith('Windows'),
+    );
+    expect(
+      windowsInfo.releaseId,
+      isNotEmpty,
+    );
+    expect(
+      windowsInfo.deviceId,
+      isNotEmpty,
+    );
+  }), skip: !Platform.isWindows);
 }