+++ packages/share_plus/share_plus/CHANGELOG.md
@@ -1,3 +1,38 @@
+## 6.1.0
+
+ - **FIX**: export XFile (#1286).
+ - **FEAT**: share XFile created using File.fromData() (#1284).
+
+## 6.0.1
+
+ - **FIX**: Increase min Flutter version to fix dartPluginClass registration (#1275).
+
+## 6.0.0
+
+> Note: This release has breaking changes.
+
+ - **FIX**: lint warnings - add missing dependency for tests (#1233).
+ - **FIX**: Show NSSharingServicePicker asynchronously on main thread (#1223).
+ - **BREAKING** **REFACTOR**: two-package federated architecture (#1238).
+
+## 5.0.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **FEAT**: Native share UI for Windows (#1158).
+
+## 4.5.3
+
+ - **CHORE**: Version tagging using melos.
+
+## 4.5.2
+
+- Update internal dependencies
+
+## 4.5.1
+
+- Update internal dependencies
+
 ## 4.5.0
 
 - iOS: Remove usage of deprecated UIApplication.keyWindow in iOS 13+