+++ packages/share_plus/share_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.2.0
+
+ - **FIX**: export XFile (#1286).
+ - **FEAT**: share XFile created using File.fromData() (#1284).
+
+## 3.1.2
+
+ - **FIX**: Increase min Flutter version to fix dartPluginClass registration (#1275).
+
+## 3.1.1
+
+- Add missing dart:ui import
+
 ## 3.1.0
 
 - Add `shareXFiles` implementations